# resource/swagger-iam.yml
openapi: 3.0.3
info:
  title: IAM Service API
  version: 1.0.0
  description: API cho microservice IAM (Identity & Access Management)

servers:
  - url: http://localhost:8080/api/v1

tags:
  - name: Auth
  - name: Users

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Đăng ký tài khoản mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignUpRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập và lấy JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /users/me:
    get:
      tags: [Users]
      summary: Lấy thông tin profile user hiện tại
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /users:
    get:
      tags: [Users]
      summary: Lấy danh sách user (admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Danh sách user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'

    post:
      tags: [Users]
      summary: Tạo user mới (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /users/{id}:
    get:
      tags: [Users]
      summary: Lấy thông tin user theo id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    put:
      tags: [Users]
      summary: Cập nhật user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    delete:
      tags: [Users]
      summary: Xóa user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted successfully

  /auth/forgot-password/request:
    post:
      tags: [Auth]
      summary: Yêu cầu OTP quên mật khẩu (gửi về email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '200':
          description: OTP đã được gửi qua email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent to your email"

  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Xác thực OTP quên mật khẩu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
              required: [email, otp]
      responses:
        '200':
          description: OTP hợp lệ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP verified successfully"

  /auth/forgot-password/reset:
    post:
      tags: [Auth]
      summary: Đặt lại mật khẩu mới sau khi xác thực OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
                newPassword:
                  type: string
                  format: password
              required: [email, otp, newPassword]
      responses:
        '200':
          description: Đặt lại mật khẩu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successfully"


components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Role:
      type: string
      enum: [ADMIN, USER, MANAGER]

    Gender:
      type: string
      enum: [MALE, FEMALE, OTHER]

    Provider:
      type: string
      enum: [LOCAL, GOOGLE, FACEBOOK, KEYCLOAK]

    UserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        dob:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          type: string
        provider:
          $ref: '#/components/schemas/Provider'
        avatarUrl:
          type: string
        allergies:
          type: array
          items:
            type: string
        eatingPreferences:
          type: array
          items:
            type: string
      required: [username, email, role, provider]

    UserResponse:
      allOf:
        - $ref: '#/components/schemas/UserRequest'
        - type: object
          properties:
            id:
              type: integer
            authId:
              type: string

    UserSignUpRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [username, email, password]

    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          format: password
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer